+++ LibCarla/source/carla/rpc/Color.h
@@ -37,8 +37,20 @@ namespace rpc {
     Color(const FColor &color)
       : Color(color.R, color.G, color.B) {}
 
+    Color(const FLinearColor &color)
+      : Color(color.R * 255.0f, color.G  * 255.0f, color.B  * 255.0f) {}
+
     operator FColor() const {
       return FColor{r, g, b};
+    }
+
+    operator FLinearColor() const {
+      return FLinearColor{
+        static_cast<float>(r)/255.0f,
+        static_cast<float>(g)/255.0f,
+        static_cast<float>(b)/255.0f,
+        1.0f
+      };
     }
 
 #endif // LIBCARLA_INCLUDED_FROM_UE4