+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Sensor/WorldObserver.h
@@ -32,7 +32,11 @@ public:
 
   /// Send a message to every connected client with the info about the given @a
   /// Episode.
-  void BroadcastTick(const UCarlaEpisode &Episode, float DeltaSeconds);
+  void BroadcastTick(
+    const UCarlaEpisode &Episode,
+    float DeltaSeconds,
+    bool MapChange,
+    bool PendingLightUpdate);
 
   /// Dummy. Required for compatibility with other sensors only.
   FTransform GetActorTransform() const