+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   * API extensions:
     - Added new methods to `Map`: `get_all_landmarks`, `get_all_landmarks_from_id` and `get_all_landmarks_of_type`
   * Added synchronization of traffic lights in sumo co-simulation
+  * Added light manager to control the lights of the map.
 
 ## CARLA 0.9.8